update Shor algorithm
