+++ src/QuAlgorithmZoo.jl
@@ -32,6 +32,7 @@ include("hamiltonian_solvers.jl")
 include("HadamardTest.jl")
 include("lin_diffEq_HHL.jl")
 include("QSVD.jl")
+include("shor.jl")
 
 
 end # module
